feat(revamp-activity-tab): implement T1 - extract run detail surface
